# Fix performance of `mesolve` with QArrays